Move class PtStory to stories folder
