+++ lib/commands/finishStory.js
@@ -1,4 +1,4 @@
-const {PtStory} = require('../../model')
+const PtStory = require('../../model/stories')
 const rebounds = require('../validation/rebounds')
 const {getState} = require('../helpers/pivotaly')
