+++ lib/commands/stopStory.js
@@ -1,4 +1,4 @@
-const {PtStory} = require('../../model')
+const PtStory = require('../../model/stories')
 const {getState} = require('../helpers/pivotaly')
 const rebounds = require('../validation/rebounds')
