+++ lib/helpers/pivotaly.js
@@ -1,5 +1,5 @@
 const {common} = require("../commands/common")
-const {PtStory} = require('../../model')
+const PtStory = require('../../model/stories')
   
 const getState = (context) => {
   const defaultState = JSON.stringify({branch: "", story: ""})
