+++ lib/validation/validators/isStoryAChore.js
@@ -1,4 +1,4 @@
-const {PtStory} = require('../../../model')
+const PtStory = require('../../../model/stories')
 
 module.exports = async (context, storyId) => {
   const storyResource = new PtStory(context, storyId)
