+++ lib/validation/validators/isStoryIDValid.js
@@ -1,4 +1,4 @@
-const {PtStory} = require('../../../model/index')
+const PtStory = require('../../../model/stories')
 const {getState} = require('../../helpers/pivotaly')
 
 exports.isStoryIDValid = async (context, storyId = null) => {
