+++ lib/validation/validators/isTokenValid.js
@@ -1,4 +1,4 @@
-const {PtStory} = require('../../../model')
+const PtStory = require('../../../model/stories')
 
 exports.isTokenValid = async (context) => {
   const storyResource = new PtStory(context, undefined)
